refactor: 다른 페이지를 만들기 위해 공통되는 컴포넌트를 사용할 수 있도록 코드 정리
