+++ src/components/App.jsx
@@ -1,5 +1,5 @@
-import Header from "components/Header";
-import Footer from "components/Footer";
+import Nav from "./Nav";
+import Footer from "./Footer";
 import "styles/reset.css";
 import "styles/color.css";
 import { Outlet } from "react-router-dom";
@@ -7,7 +7,7 @@ import { Outlet } from "react-router-dom";
 function App() {
   return (
     <>
-      <Header />
+      <Nav />
       <div>
         <Outlet />
       </div>
