+++ src/components/Nav/Nav.css
@@ -1,11 +1,3 @@
-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  background-color: #edf7ff;
-}
-
 nav {
   display: flex;
   justify-content: space-between;
