+++ src/pages/SharedPage/SharedPage.jsx
@@ -1,13 +1,30 @@
 import CardList from "components/CardList";
 import SearchBar from "components/SearchBar";
 import "./SharedPage.css";
+import FolderInfo from "components/FolderInfo";
+import { useEffect, useState } from "react";
+import { getFolder } from "api";
 
 function SharedPage() {
+  const [folder, setFolder] = useState({});
+
+  const handleLoadFolder = async () => {
+    const result = await getFolder();
+    setFolder(result);
+  };
+
+  useEffect(() => {
+    handleLoadFolder();
+  }, []);
+
   return (
-    <main>
-      <SearchBar />
-      <CardList />
-    </main>
+    <>
+      <FolderInfo data={folder} />
+      <main>
+        <SearchBar />
+        <CardList />
+      </main>
+    </>
   );
 }
